+++ canonicalwebteam/discourse/__init__.py
@@ -5,7 +5,7 @@ from canonicalwebteam.discourse.app import (  # noqa
     Category,  # noqa
     Events,  # noqa
 )
-from canonicalwebteam.discourse.cache import ResponseCache  # noqa
+from canonicalwebteam.discourse.db_cache import DBResponseCache  # noqa
 from canonicalwebteam.discourse.exceptions import RateLimitedError  # noqa
 from canonicalwebteam.discourse.models import DiscourseAPI  # noqa
 from canonicalwebteam.discourse.parsers import (  # noqa